requirements.txt
python-binance==1.0.19
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
schedule==1.2.0
openai==1.3.0
transformers==4.34.0
requests==2.31.0
ccxt==4.1.13
ta-lib==0.4.28
scikit-learn==1.3.0
tensorflow==2.13.0
newsapi-python==0.2.7 